push remainder of changes

+++ express-back-end/db/schema/01_tweeter.sql
@@ -3,7 +3,7 @@
 DROP TABLE IF EXISTS tweeters CASCADE;
 CREATE TABLE tweeters (
   id SERIAL PRIMARY KEY NOT NULL,
-  tweeter_id INT NOT NULL,
+  tweeter_id BIGINT NOT NULL,
   screen_name VARCHAR(255) NOT NULL,
   location VARCHAR(255),
   profile_image_url VARCHAR(255)
